+++ lib/rubocop/cop/metrics/block_nesting.rb
@@ -2,7 +2,7 @@
 
 module RuboCop
   module Cop
-    module Style
+    module Metrics
       # This cop checks for excessive nesting of conditional and looping
       # constructs. Despite the cop's name, blocks are not considered as an
       # extra level of nesting.