+++ spec/rubocop/cop/metrics/perceived_complexity_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
+describe RuboCop::Cop::Metrics::PerceivedComplexity, :config do
   subject(:cop) { described_class.new(config) }
 
   context 'when Max is 1' do
@@ -33,7 +33,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo if some_condition',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
       expect(cop.highlights).to eq(['def'])
       expect(cop.config_to_allow_offenses).to eq('Max' => 2)
     end
@@ -43,10 +43,10 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo unless some_condition',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
-    it 'registers an offense for an elsif block' do
+    it 'registers an offense for elsif and else blocks' do
       inspect_source(cop, ['def method_name',
                            '  if first_condition then',
                            '    call_foo',
@@ -57,7 +57,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  end',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [3/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [4/1]'])
     end
 
     it 'registers an offense for a ternary operator' do
@@ -65,7 +65,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  value = some_condition ? 1 : 2',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for a while block' do
@@ -75,7 +75,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  end',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for an until block' do
@@ -85,7 +85,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  end',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for a for block' do
@@ -95,7 +95,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  end',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for a rescue block' do
@@ -107,20 +107,38 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  end',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for a case/when block' do
       inspect_source(cop, ['def method_name',
                            '  case value',
-                           '  when 1',
+                           '  when 1 then call_foo_1',
+                           '  when 2 then call_foo_2',
+                           '  when 3 then call_foo_3',
+                           '  when 4 then call_foo_4',
+                           '  end',
+                           'end'])
+      # The `case` node plus the first `when` score one complexity point
+      # together. The other `when` nodes get 0.2 complexity points.
+      expect(cop.messages)
+        .to eq(['Perceived complexity for method_name is too high. [3/1]'])
+    end
+
+    it 'registers an offense for a case/when block without an expression ' \
+       'after case' do
+      inspect_source(cop, ['def method_name',
+                           '  case',
+                           '  when value == 1',
                            '    call_foo',
-                           '  when 2',
+                           '  when value == 2',
                            '    call_bar',
                            '  end',
                            'end'])
+      # Here, the `case` node doesn't count, but each when scores one
+      # complexity point.
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [3/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [3/1]'])
     end
 
     it 'registers an offense for &&' do
@@ -128,7 +146,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo && call_bar',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for and' do
@@ -136,7 +154,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo and call_bar',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for ||' do
@@ -144,7 +162,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo || call_bar',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'registers an offense for or' do
@@ -152,7 +170,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo or call_bar',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [2/1]'])
     end
 
     it 'deals with nested if blocks containing && and ||' do
@@ -164,7 +182,7 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                       '  end',
                       'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [6/1]'])
+        .to eq(['Perceived complexity for method_name is too high. [6/1]'])
     end
 
     it 'counts only a single method' do
@@ -176,8 +194,8 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
                            '  call_foo if some_condition',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name_1 is too high. [2/1]',
-                'Cyclomatic complexity for method_name_2 is too high. [2/1]'])
+        .to eq(['Perceived complexity for method_name_1 is too high. [2/1]',
+                'Perceived complexity for method_name_2 is too high. [2/1]'])
     end
   end
 
@@ -185,20 +203,20 @@ describe RuboCop::Cop::Style::CyclomaticComplexity, :config do
     let(:cop_config) { { 'Max' => 2 } }
 
     it 'counts stupid nested if and else blocks' do
-      inspect_source(cop, ['def method_name',
-                           '  if first_condition then',
+      inspect_source(cop, ['def method_name',                   # 1
+                           '  if first_condition then',         # 2
                            '    call_foo',
-                           '  else',
-                           '    if second_condition then',
+                           '  else',                            # 3
+                           '    if second_condition then',      # 4
                            '      call_bar',
-                           '    else',
-                           '      call_bam if third_condition',
+                           '    else',                          # 5
+                           '      call_bam if third_condition', # 6
                            '    end',
-                           '    call_baz if fourth_condition',
+                           '    call_baz if fourth_condition',  # 7
                            '  end',
                            'end'])
       expect(cop.messages)
-        .to eq(['Cyclomatic complexity for method_name is too high. [5/2]'])
+        .to eq(['Perceived complexity for method_name is too high. [7/2]'])
     end
   end
 end